+++ carp_core/test/carp_core_kotlin_rpc_test.dart
@@ -498,7 +498,7 @@ void main() {
 
       var request = SetParticipantData(
         'c9cc5317-48da-45f2-958e-58bc07f34681',
-        {SexCustomInput.SEX_INPUT_TYPE_NAME: SexCustomInput(Sex.Male)},
+        {SexInput.type: SexInput(Sex.Male)},
         'Participant',
       );